started moving into separate files and classes
